requirements.txt
certifi
fitz
flask
openai
streamlit
pymupdf
flask-cors
bs4
tldextract
python-dotenv